// src/ledger/interfaces/ledger.interface.ts
export enum Account {
  CHECK_ASSETS = 'Check Assets',
  ACCOUNTS_RECEIVABLE = 'Accounts Receivable',
  LIABILITY = 'Liability',
  FEE_INCOME = 'Fee Income',
  INCOME = 'Income',
  CREDITS_ISSUED = 'Credits Issued',
  ASSET = 'Asset',
}

export enum SubAccount {
  RENT = 'Rent',
  SECURITY_DEPOSITS = 'Security Deposits',
  PREPAID_INSURANCE_LIABILITIES = 'Prepaid Insurance Liabilities',
  LATE_FEES = 'Late Fees',
  ADMIN_FEE = 'Admin Fee',
  MERCHANDISE_INCOME = 'Merchandise Income',
  RENT_WAIVED = 'Rent Waived',
  LATE_FEES_WAIVED = 'Late Fees Waived',
  INSURANCE_WAIVED = 'Insurance Waived',
  ACCOUNTS_RECEIVABLE = 'Accounts Receivable',
  ADMINISTRATIVE_FEES = 'Administrative Fees',
  RENTAL_INCOME = 'Rental Income',
  INSURANCE_INCOME = 'Insurance Income',
  INSURANCE = 'Insurance',
  FEES = 'FEES',
}